requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.11.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
